# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hotel Guest Check-in API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /checkin/{token}:
    get:
      operationId: getCheckinSession
      tags: [guests]
      summary: Get booking info for guest check-in form (public)
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Booking info for guest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckinSession"
        "404":
          description: Invalid or expired token
  /checkin/{token}/guests:
    post:
      operationId: submitGuestCheckin
      tags: [guests]
      summary: Submit guest check-in information
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestCheckinInput"
      responses:
        "201":
          description: Guest check-in submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestRecord"
        "400":
          description: Invalid data
        "404":
          description: Invalid token
  /guests/{id}/upload:
    post:
      operationId: uploadGuestDocument
      tags: [guests]
      summary: Upload ID document or signature for a guest
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentUploadInput"
      responses:
        "200":
          description: Upload result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestRecord"
        "404":
          description: Guest not found
...
  /guests/search:
    get:
      operationId: searchGuests
      tags: [guests]
      summary: Search guests across all bookings
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Matching guests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GuestSearchResult"
...
components:
  schemas:
...
    CheckinSession:
      type: object
      required: [bookingId, bookingRef, guestName, roomNumber, checkInDate, checkOutDate, numberOfGuests, guestsSubmitted]
      properties:
        bookingId:
          type: integer
        bookingRef:
          type: string
        guestName:
          type: string
        roomNumber:
          type: string
        checkInDate:
          type: string
        checkOutDate:
          type: string
        numberOfGuests:
          type: integer
        guestsSubmitted:
          type: integer
    GuestRecord:
      type: object
      required: [id, bookingId, fullName, phone, email, address, nationality, idType, isForeignNational, createdAt]
      properties:
        id:
          type: integer
        bookingId:
          type: integer
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        nationality:
          type: string
        idType:
          type: string
          enum: [aadhaar, passport, driving_licence, voter_id, other]
        idNumber:
          type: ["string", "null"]
        idImageUrl:
          type: ["string", "null"]
        signatureUrl:
          type: ["string", "null"]
        isForeignNational:
          type: boolean
        numberOfGuests:
          type: ["integer", "null"]
        visaNumber:
          type: ["string", "null"]
        visaType:
          type: ["string", "null"]
        visaExpiry:
          type: ["string", "null"]
        portOfEntry:
          type: ["string", "null"]
        arrivalDate:
          type: ["string", "null"]
        passportNumber:
          type: ["string", "null"]
        passportExpiry:
          type: ["string", "null"]
        createdAt:
          type: string
    GuestCheckinInput:
      type: object
      required: [fullName, phone, email, address, nationality, idType]
      properties:
        fullName:
          type: string
          minLength: 1
        phone:
          type: string
          minLength: 1
        email:
          type: string
        address:
          type: string
          minLength: 1
        nationality:
          type: string
          minLength: 1
        idType:
          type: string
          enum: [aadhaar, passport, driving_licence, voter_id, other]
        idNumber:
          type: string
        isForeignNational:
          type: boolean
        numberOfGuests:
          type: integer
        visaNumber:
          type: string
        visaType:
          type: string
        visaExpiry:
          type: string
        portOfEntry:
          type: string
        arrivalDate:
          type: string
        passportNumber:
          type: string
        passportExpiry:
          type: string
    DocumentUploadInput:
      type: object
      required: [type, dataUrl, fileName]
      properties:
        type:
          type: string
          enum: [id_document, signature, passport]
        dataUrl:
          type: string
        fileName:
          type: string
    GuestSearchResult:
      type: object
      required: [id, bookingId, bookingRef, fullName, phone, email, nationality, idType, isForeignNational, createdAt]
      properties:
        id:
          type: integer
        bookingId:
          type: integer
        bookingRef:
          type: string
        roomNumber:
          type: string
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
        nationality:
          type: string
        idType:
          type: string
        isForeignNational:
          type: boolean
        createdAt:
          type: string
...