# lib/api-zod/src/generated/types/documentUploadInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Hotel Guest Check-in API
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentUploadInputType } from './documentUploadInputType';

export interface DocumentUploadInput {
  type: DocumentUploadInputType;
  dataUrl: string;
  fileName: string;
}